Modified login form to use HtmlHelper, including adding custom templates.
These custom templates will also be used elsewhere in the app when a
form is required. However for now they are only supported in the login
form. The custom templates also only cover what was required for that
login form. In the future, as more form elements are required, the
templates will need to be added to.