Create a proper administration menu, linking to real controllers + actions.
This has been done in such a way that adding a new menu item requires
editing a PHP array and adding another element. This is in contrast to
manually editing HTML. I prefer this approach, because I don't need to
know what the actual HTML looks like for the menu in order to add more
items or make changes, which is helpful. It will prevent errors in the
future whereby someone may add another menu item, but accidentally style
it in a manner inconsistent with the other menu items.