Support the notion of a "demo site", which allows anyone to login.
The login form will pre-populate itself with "root@example.com" and
"demo password". These will always log you in as the first user in demo
mode, regardless of whether someone changes the users email address.

If either of those values are changed though, then it will still fail
the login check, in order to demonstrate that the login feature works as
expected.