Show all articles with a tag in the public template.
Uses the same article "Snippet" or summary as the home page. To ensure
that this code was not duplicated, the code responsible for rendering
the title + snippet + post date was extracted into an "Element". This
means that in the future, if we were, for example, to add an "Author" to
each article on the home page, it will automatically be added to all
articles in the public "Tags" page too.