Don't escape HTML tags in $article->body as the TinyMCE editor has added HTML.
THIS IS A BIG SECURITY PROBLEM, and will be addressed in the following
commit. Indeed I wouldn't usually even consider putting this type of
commit in git, but it is here to explain the steps between showing just
plain text from the article, to showing HTML from the article in an
UNSAFE manner, then showing it safely. I plan to use a 3rd party library
called HTMLPurifier to ensure that only HTML which is considered safe is
allowed to be included in the article body.