Correctly redirect to admin dashboard after logging in.
Pass through a "?redirect=/admin" argument to the login form from the
home page. Also make sure that when the form is submitted, it includes
"?redirect=..." in the submission URL. This will get picked up by the
`UsersController` when it performs the login, by virtue of the
`$this->Auth->redirectUrl()` method.