Simplify the code for the multi-select form control.
Previously, this added a new form template, which adds a "chosen" class
to the widget. The JavaScript responsible for initializing the Chosen
library would look for select widgets with that class (i.e.
`select.chosen`). Now it blindly applies the chosen widget to all select
lists in the admin dashboard regardless of class name.