Baked the ArticlesController to include better validation and comments.
Specifically, the slug should be unique, and Cake Bake knows how to deal
with this well. The version of ArticlesTable I had prior to this was
hand crafted by following the CMS tutorial in the cookbook, which missed
out on this detail.

This also allows the slug of existing articles to be modified via the
edit article form.

Finally, it adds a "View" button to the edit article page.