Add helper script to update articles SQL file, for easier import.
Before, I had to manually add the `INSERT INTO articles` code for each
article I wanted to keep with the default schema. That was prettty
tedious, so this adds a script which dumps the contents of the articles
(including some fake "article views" data) ready for import elsewhere.

It is intentionally kept separate from the "database.sql" file, because
that is hand crafted and should be easy to edit, whereas the intention
is that this article content goes into an automatically generated file
that can be overridden whenever you please.