Minor improvements to the style of articles when displayed in public template.
Code is shown with a background and a little bit of padding now (whether
it is inline via a `<code>` tag, or a block such as `<pre>`). Also strip
HTML tags before showing article snippets. This does a few things, it
prevents things like headings and images being displayed in an article
list, and it also stops broken HTML (e.g. trimming the text halfway
between "<h1>...." before getting to "</h1>".