Connected the search box with the database, show search results.
Currently searches article titles + body for text that matches
the query. Known issues icnlude: It doesn't strpi HTML in the database
before searching, so technically you could search for "href" and it would
probably come back with results. It also doesn't search fo rmultiple
words, it only matches exact phrases. So if I was to search for "first
broken", and there was an article containing "The first record was
broken", then it wont match, because it is looking for the exact phrase
"first broken".